+++ BambooTracker/gui/instrument_editor/instrument_editor_ssg_form.cpp
@@ -5,6 +5,7 @@
 #include <stdexcept>
 #include "gui/event_guard.hpp"
 #include "pitch_converter.hpp"
+#include "command_sequence.hpp"
 #include "misc.hpp"
 
 InstrumentEditorSSGForm::InstrumentEditorSSGForm(int num, QWidget *parent) :
@@ -139,16 +140,22 @@ InstrumentEditorSSGForm::InstrumentEditorSSGForm(int num, QWidget *parent) :
 	}
 	ui->envEditor->setMultipleReleaseState(true);
 
-	ui->hardFreqSpinBox->setSuffix(
-				QString(" (0x") + QString("%1 | ").arg(ui->hardFreqSpinBox->value(), 3, 16, QChar('0')).toUpper()
-				 + QString("%1Hz)").arg(QString::number(7800.0 / ui->hardFreqSpinBox->value(), 'f', 4)));
-
 	QObject::connect(ui->envEditor, &VisualizedInstrumentMacroEditor::sequenceCommandAdded,
 					 this, [&](int row, int col) {
 		if (!isIgnoreEvent_) {
 			if (row >= 16) {	// Set hard frequency
-				ui->envEditor->setText(col, QString::number(ui->hardFreqSpinBox->value()));
-				ui->envEditor->setData(col, ui->hardFreqSpinBox->value());
+				if (ui->hardFreqButtonGroup->checkedButton() == ui->hardFreqRawRadioButton) {
+					ui->envEditor->setText(col, QString::number(ui->hardFreqRawSpinBox->value()));
+					ui->envEditor->setData(col, ui->hardFreqRawSpinBox->value());
+				}
+				else {
+					ui->envEditor->setText(col, QString::number(ui->hardFreqToneSpinBox->value()) + "/"
+										   + QString::number(ui->hardFreqHardSpinBox->value()));
+
+					ui->envEditor->setData(col, CommandSequenceUnit::ratio2data(
+											   ui->hardFreqToneSpinBox->value(),
+											   ui->hardFreqHardSpinBox->value()));
+				}
 			}
 			bt_.lock()->addEnvelopeSSGSequenceCommand(
 						ui->envNumSpinBox->value(), row, ui->envEditor->getSequenceDataAt(col));
@@ -168,8 +175,17 @@ InstrumentEditorSSGForm::InstrumentEditorSSGForm(int num, QWidget *parent) :
 					 this, [&](int row, int col) {
 		if (!isIgnoreEvent_) {
 			if (row >= 16) {	// Set hard frequency
-				ui->envEditor->setText(col, QString::number(ui->hardFreqSpinBox->value()));
-				ui->envEditor->setData(col, ui->hardFreqSpinBox->value());
+				if (ui->hardFreqButtonGroup->checkedButton() == ui->hardFreqRawRadioButton) {
+					ui->envEditor->setText(col, QString::number(ui->hardFreqRawSpinBox->value()));
+					ui->envEditor->setData(col, ui->hardFreqRawSpinBox->value());
+				}
+				else {
+					ui->envEditor->setText(col, QString::number(ui->hardFreqToneSpinBox->value()) + "/"
+										   + QString::number(ui->hardFreqHardSpinBox->value()));
+					ui->envEditor->setData(col, CommandSequenceUnit::ratio2data(
+											   ui->hardFreqToneSpinBox->value(),
+											   ui->hardFreqHardSpinBox->value()));
+				}
 			}
 			bt_.lock()->setEnvelopeSSGSequenceCommand(
 						ui->envNumSpinBox->value(), col, row, ui->envEditor->getSequenceDataAt(col));
@@ -696,7 +712,13 @@ void InstrumentEditorSSGForm::setInstrumentEnvelopeParameters()
 	for (auto& com : instSSG->getEnvelopeSequence()) {
 		QString str = "";
 		if (com.type >= 16) {
-			str = QString("%1").arg(com.data);
+			if (CommandSequenceUnit::isRatioData(com.data)) {
+				auto ratio = CommandSequenceUnit::data2ratio(com.data);
+				str = QString::number(ratio.first) + "/" + QString::number(ratio.second);
+			}
+			else {
+				str = QString::number(com.data);
+			}
 		}
 		ui->envEditor->addSequenceCommand(com.type, str, com.data);
 	}
@@ -760,10 +782,10 @@ void InstrumentEditorSSGForm::on_envNumSpinBox_valueChanged(int arg1)
 	onEnvelopeNumberChanged();
 }
 
-void InstrumentEditorSSGForm::on_hardFreqSpinBox_valueChanged(int arg1)
+void InstrumentEditorSSGForm::on_hardFreqRawSpinBox_valueChanged(int arg1)
 {
-	ui->hardFreqSpinBox->setSuffix(
-				QString(" (0x") + QString("%1 | ").arg(arg1, 3, 16, QChar('0')).toUpper()
+	ui->hardFreqRawSpinBox->setSuffix(
+				QString(" (0x") + QString("%1 | ").arg(arg1, 4, 16, QChar('0')).toUpper()
 				+ QString("%1Hz").arg(arg1 ? QString::number(7800.0 / arg1, 'f', 4) : "-"));
 }