+++ BambooTracker/io/instrument_io.cpp
@@ -384,7 +384,7 @@ void InstrumentIO::saveInstrument(std::string path, std::weak_ptr<InstrumentsMan
 			ctr.appendUint16(static_cast<uint16_t>(seq.size()));
 			for (auto& com : seq) {
 				ctr.appendUint16(static_cast<uint16_t>(com.type));
-				ctr.appendUint16(static_cast<uint16_t>(com.data));
+				ctr.appendInt32(static_cast<int32_t>(com.data));
 			}
 			auto loop = instMan.lock()->getWaveFormSSGLoops(wfNum);
 			ctr.appendUint16(static_cast<uint16_t>(loop.size()));
@@ -465,7 +465,7 @@ void InstrumentIO::saveInstrument(std::string path, std::weak_ptr<InstrumentsMan
 			ctr.appendUint16(static_cast<uint16_t>(seq.size()));
 			for (auto& com : seq) {
 				ctr.appendUint16(static_cast<uint16_t>(com.type));
-				ctr.appendUint16(static_cast<uint16_t>(com.data));
+				ctr.appendInt32(static_cast<int32_t>(com.data));
 			}
 			auto loop = instMan.lock()->getEnvelopeSSGLoops(envNum);
 			ctr.appendUint16(static_cast<uint16_t>(loop.size()));
@@ -1571,10 +1571,15 @@ AbstractInstrument* InstrumentIO::loadBTIFile(std::string path,
 							if (data == 3) data = static_cast<int>(SSGWaveFormType::SQM_TRIANGLE);
 							else if (data == 4) data = static_cast<int>(SSGWaveFormType::SQM_SAW);
 						}
-						uint16_t subdata = ctr.readUint16(csr);
-						csr += 2;
-						if (fileVersion < Version::toBCD(1, 2, 0)) {
-							if (subdata != 0xffff)
+						int32_t subdata;
+						if (fileVersion >= Version::toBCD(1, 2, 0)) {
+							subdata = ctr.readInt32(csr);
+							csr += 4;
+						}
+						else {
+							subdata = ctr.readUint16(csr);
+							csr += 2;
+							if (subdata != -1)
 								subdata = PitchConverter::getPitchSSGSquare(subdata);
 						}
 						if (l == 0)
@@ -1704,8 +1709,15 @@ AbstractInstrument* InstrumentIO::loadBTIFile(std::string path,
 					for (uint16_t l = 0; l < seqLen; ++l) {
 						uint16_t data = ctr.readUint16(csr);
 						csr += 2;
-						uint16_t subdata = ctr.readUint16(csr);
-						csr += 2;
+						int32_t subdata;
+						if (fileVersion >= Version::toBCD(1, 2, 0)) {
+							subdata = ctr.readInt32(csr);
+							csr += 4;
+						}
+						else {
+							subdata = ctr.readUint16(csr);
+							csr += 2;
+						}
 						if (l == 0)
 							instMan.lock()->setEnvelopeSSGSequenceCommand(idx, 0, data, subdata);
 						else